server/requirements.txt
autopep8==1.2.2
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
boto==2.39.0
certifi==2015.11.20.1
coverage==4.0.3
Django==1.8.9
django-boto==0.3.9
django-braces==1.8.1
django-cors-headers==1.1.0
django-earthdistance==1
django-extensions==1.6.1
django-filter==0.12.0
django-jenkins==0.18.1
django-kronos==0.9
django-oauth-toolkit==0.10.0
django-rest-framework-social-oauth2==1.0.2
django-rest-swagger==0.3.5
django-ses==0.7.0
django-silk==0.5.7
django-storages==1.1.8
django-uuidfield==0.5.0
djangorestframework==3.3.2
ecdsa==0.13
elasticsearch==2.2.0
functools32==3.2.3.post2
futures==3.0.4
geopy==1.11.0
gevent==1.0.2
gevent-socketio==0.3.6
greenlet==0.4.9
gunicorn==19.4.5
httplib2==0.9.2
image==1.4.4
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
newrelic==2.60.0.46
oauthlib==1.0.3
paramiko==1.16.0
pep8==1.7.0
Pillow==3.1.0
poster==0.8.1
psutil==3.4.2
psycopg2==2.6.1
pycrypto==2.6.1
Pygments==2.1.3
PyJWT==1.4.0
python-crontab==2.0.1
python-dateutil==2.4.2
python-google-places==1.1.0
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2016.3
PyYAML==3.11
redis==2.10.5
requests==2.9.1
requests-aws4auth==0.8
requests-oauthlib==0.6.0
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
sleekxmpp==1.3.1
sockjs-tornado==1.0.2
sphinx-me==0.3
sqlparse==0.1.19
strict-rfc3339==0.6
tinys3==0.1.11
tornado==4.3
tornado-redis==2.4.18
txaio==2.2.1
urllib3==1.14
uWSGI==2.0.12
zope.interface==4.1.3
